Rename app identifiers for public prototype
