+++ android/app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:429989696553:android:2b43afa49913cccede3838",
         "android_client_info": {
-          "package_name": "com.example.flutter_application_1"
+          "package_name": "com.parkinsum.companion"
         }
       },
       "oauth_client": [],
