+++ android/app/src/main/kotlin/com/parkinsum/companion/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.flutter_application_1
+package com.parkinsum.companion
 
 import io.flutter.embedding.android.FlutterActivity
