+++ macos/Runner/GoogleService-Info.plist
@@ -9,7 +9,7 @@
 	<key>PLIST_VERSION</key>
 	<string>1</string>
 	<key>BUNDLE_ID</key>
-	<string>com.example.flutterApplication1</string>
+	<string>com.parkinsum.companion</string>
 	<key>PROJECT_ID</key>
 	<string>parkinsum-companion</string>
 	<key>STORAGE_BUCKET</key>
