+++ windows/runner/Runner.rc
@@ -89,13 +89,13 @@ BEGIN
     BEGIN
         BLOCK "040904e4"
         BEGIN
-            VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "flutter_application_1" "\0"
+            VALUE "CompanyName", "ParkinSUM" "\0"
+            VALUE "FileDescription", "ParkinSUM Companion" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "flutter_application_1" "\0"
-            VALUE "LegalCopyright", "Copyright (C) 2026 com.example. All rights reserved." "\0"
-            VALUE "OriginalFilename", "flutter_application_1.exe" "\0"
-            VALUE "ProductName", "flutter_application_1" "\0"
+            VALUE "InternalName", "parkinsum_companion" "\0"
+            VALUE "LegalCopyright", "Copyright (C) 2026 ParkinSUM. All rights reserved." "\0"
+            VALUE "OriginalFilename", "parkinsum_companion.exe" "\0"
+            VALUE "ProductName", "ParkinSUM Companion" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
